Move QueryScheduleAutoMode to a new module
