+++ src/nextline_schedule/schema/auto/__init__.py
@@ -1,5 +1,7 @@
 __all__ = [
     'MutationScheduleAutoMode',
+    'QueryScheduleAutoMode',
 ]
 
 from .mutation import MutationScheduleAutoMode
+from .query import QueryScheduleAutoMode
