+++ src/nextline_schedule/schema/query.py
@@ -2,21 +2,10 @@ import strawberry
 from strawberry.types import Info
 
 from nextline_schedule import __version__
-from nextline_schedule.auto import AutoMode
 from nextline_schedule.scheduler import RequestStatement
 
+from .auto import QueryScheduleAutoMode
 from .queue import QueryScheduleQueue
-
-
-def query_auto_mode_state(info: Info) -> str:
-    auto_mode = info.context['schedule']['auto_mode']
-    assert isinstance(auto_mode, AutoMode)
-    return auto_mode.state
-
-
-@strawberry.type
-class QueryScheduleAutoMode:
-    state: str = strawberry.field(resolver=query_auto_mode_state)
 
 
 def query_scheduler_api_url(info: Info) -> str:
